Update Dockerfile complete build
This PR is needed to run a complete Java Project build inside the dockerfile. Our Dockerfiles will no longer be able to count on the Toolchain or external environment to always compile the .jar file. We are updating this dockerfile to run a 2-stage Docker build, where the first stage downloads dependencies, and compiles the .jar, and then the second stage copies it over from the first container, and the first container is destroyed. This change is to work with the External OpenToolchain Toolchain templates that DevX is transitioning to.